Remove deprecated config onBrokenMarkdownLinks (#297)
`siteConfig.onBrokenMarkdownLinks` is deprecated:

[WARNING] The `siteConfig.onBrokenMarkdownLinks` config option is
deprecated and will be removed in Docusaurus v4.
Please migrate and move this option to
`siteConfig.markdown.hooks.onBrokenMarkdownLinks` instead.

In the beginning, the default value is `warn`, so we don't need to set
this option.


https://docusaurus.io/docs/api/docusaurus-config#hooks.onBrokenMarkdownLinks

Note: If we set this option explicitly, we can as following:

    markdown: {
      hooks: {
        onBrokenMarkdownLinks: 'warn',
      },
    },